LayoutTests/http/tests/security/cannot-read-cssrules-expected.txt

CONSOLE MESSAGE: Origin http://127.0.0.1:8000 is not allowed by Access-Control-Allow-Origin. Status code: 200
This test whether a script can read the rules from a cross-origin style sheet. For more information on why we block this, please see https://bugs.webkit.org/show_bug.cgi?id=20527.

Test begins.
== Cross-Origin resource, no-cors mode ==
exception thrown for cssRules: SecurityError: Not allowed to access cross-origin stylesheet
exception thrown for rules: SecurityError: Not allowed to access cross-origin stylesheet
== Same-Origin resource, no-cors mode ==
cssRules: [object CSSRuleList]
cssRules length: 1
rules: [object CSSRuleList]
rules length: 1
== Cross-Origin resource, cors mode, failing cors check ==
exception thrown for cssRules: SecurityError: Not allowed to access cross-origin stylesheet
exception thrown for rules: SecurityError: Not allowed to access cross-origin stylesheet
== Cross-Origin resource, cors mode, successful cors check ==
cssRules: [object CSSRuleList]
cssRules length: 1
rules: [object CSSRuleList]
rules length: 1
Test ends.
